+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Changes and Fixes
 
+- Module [AHAB](./docs/ahab/Ahab.md)
+    - Fixed reset behavior on loading files or examples
+    - Fixed wrong MIDI channel on MIDI output messages
+    - Fixed various minor issues
 - Module [MB](./docs/trial-and-error/Mb.md)
     - Added sorting option "Newest" to *v2 mod* browser
     - Improved context menus for tags (#461)
@@ -12,6 +16,7 @@
     - Fixed incorrect reset behavior
 - Module [MIDI-MON](./docs/midi/MidiMon.md)
     - Splitted display option for 14bit CC and RPN/NRPN messages
+    - Increased buffered message count from 512 to 4096
 - Module [SIREN](./docs/siren/Siren.md)
     - Added support for multi-channel files on Re-pitch and Crossfade-loop
     - Ignore files starting with "." (#460)